Added Loading Icon, Fixed Scrolling, Changed Proof Display
Added a loading icon whenever Bunches or Pokemon are being loaded. Fixed scrolling not working properly on Firefox and made it so if clicked on a Pokemon the Viewing Area would hide upon turning on, the Selection Area would scroll to display it at the new bottom.  Fixed a text size bug for bunches on mobile, made it able to see proof without having to open a new tab (in beta)