Added Ribbons, UUID, Refresh Prompts, Standard View
Added Ribbons for pokemon, UUID for use in the URL instead of the search ID, Refresh Prompts to force a version change. Removed the option to have a smaller view for the pokemon as with recent changes it didn't work properly when clicked anyways. Added cursor for clicks, filter options, bug fixes and slight changes for a better experience.